--- python_modules/libraries/dagster-fivetran/dagster_fivetran/experimental/__init__.py
@@ -1,1 +1,0 @@
-from dagster_fivetran.experimental.resources import FivetranWorkspace as FivetranWorkspace
